Documentation update
Documentation v0.1

+++ server/API/accounts_manager.py
@@ -62,14 +62,15 @@ class AccountManager:
             info = request.json
             users:dict = globals()["users"]
             if info.get("name") in users:
-                con = sqlite3.connect("Databases\\Users.sqlite")
-                cur = con.cursor()
-                res = cur.execute(f"DELETE FROM 'users' WHERE name = '{info.get('name')}'")
-                con.commit()
-                User().update_token()
-                log_message = f"Фccount with name: {info.get('name')} was deleted"
-                loger.info("URAccount", log_message)
-                return jsonify({"status": True, "info": log_message}), 200
+                if users[info.get("name")]["role"] not in {"SuperAdmin", "System"}:
+                    con = sqlite3.connect("Databases\\Users.sqlite")
+                    cur = con.cursor()
+                    res = cur.execute(f"DELETE FROM 'users' WHERE name = '{info.get('name')}'")
+                    con.commit()
+                    User().update_token()
+                    log_message = f"Account with name: {info.get('name')} was deleted"
+                    loger.info("URAccount", log_message)
+                    return jsonify({"status": True, "info": log_message}), 200
             else:
                 return jsonify({"status": False, "info": "No such account exists"}), 400
             
@@ -93,16 +94,20 @@ class AccountManager:
             if info.get("server_token") == server_token.reg_token:
                 User().update_token()
                 if info.get("name") in users:
-                    if users[info.get("name")]["password"] == info.get("password"):
-                        role = users[info.get('name')]['role']
-                        token = users[info.get('name')]['token']
-                        return jsonify({"status": True, "info": "User found", "role": role, "token": token}), 200
+                    if users[info.get("name")]["role"] not in {"System"}:
+                        if users[info.get("name")]["password"] == info.get("password"):
+                            role = users[info.get('name')]['role']
+                            token = users[info.get('name')]['token']
+                            return jsonify({"status": True, "info": "User found", "role": role, "token": token}), 200
+                        else:
+                            loger.error("URAccount", f"Password incorrect")
+                            return jsonify({"status": False, "info": "Password incorrect"}), 400
                     else:
-                        loger.error("URAccount", f"Password incorrect")
-                        return jsonify({"status": False, "info": "Password incorrect"}), 400
+                        loger.error("URAccount", f"Account data with the System role cannot be transferred") # TODO: add ip address user to log
+                        return jsonify({"status": False, "info": "Account data with the System role cannot be transferred"}), 400
                 else:
                     loger.error("URAccount", f"Name not in users")
-                    return jsonify({"status": False, "info": "Name not in users"}), 400
+                    return jsonify({"status": False, "info": "Name not in users"}), 404
             else:
                 loger.error("URAccount", f"Server token incorrect")
                 return jsonify({"status": False, "info": "Server token incorrect"}), 400
@@ -112,28 +117,44 @@ class AccountManager:
         @access.check_user(user_role="SuperAdmin", loger_module=self.loger_module)
         def change_password():
             info = request.json
-            con = sqlite3.connect("Databases\\Users.sqlite")
-            cur = con.cursor()
-            cur.execute(f"UPDATE users SET password = '{info.get('password')}' WHERE name = '{info.get('name')}'")
-            con.commit()
-            con.close()
-            User().update_token()
-            log_message = f"Password was changed for account with name: {info.get('name')}"
-            loger.info("URAccount", )
-            return jsonify({"status": True, "info": log_message}), 200
-
+            users:dict = globals()["users"]
+            if info.get("name") in users:
+                if info.get("name") != "":
+                    con = sqlite3.connect("Databases\\Users.sqlite")
+                    cur = con.cursor()
+                    cur.execute(f"UPDATE users SET password = '{info.get('password')}' WHERE name = '{info.get('name')}'")
+                    con.commit()
+                    con.close()
+                    User().update_token()
+                    log_message = f"Password was changed for account with name: {info.get('name')}"
+                    loger.info("URAccount", )
+                    return jsonify({"status": True, "info": log_message}), 200
+                loger.error("URAccount", f"You try change password system account")
+                return jsonify({"status": False, "info": "You try change password system account"}), 400
+            else:
+                loger.error("URAccount", f"Name not in users")
+                return jsonify({"status": False, "info": "Name not in users"}), 404
         # get user token
         @app.route("/GetToken", methods=['POST'])
         @access.check_user(user_role="SuperAdmin", loger_module=self.loger_module)
         def get_user_token():
             info = request.json
-            con = sqlite3.connect("Databases\\Users.sqlite")
-            cur = con.cursor()
-            cur.execute(f"SELECT token FROM users WHERE name = '{info.get('name')}, password = '{info.get('password')}'")
-            token = cur.fetchone()
-            con.commit()
-            con.close()
-            return jsonify({"status": True, "info": "", "token": token}), 200
+            users:dict = globals()["users"]
+            if info.get("name") in users:
+                if info.get("name") != "":
+                    con = sqlite3.connect("Databases\\Users.sqlite")
+                    cur = con.cursor()
+                    cur.execute(f"SELECT token FROM users WHERE name = '{info.get('name')}, password = '{info.get('password')}'")
+                    token = cur.fetchone()
+                    con.commit()
+                    con.close()
+                    return jsonify({"status": True, "info": "", "token": token}), 200
+                else:
+                    loger.error("URAccount", f"You try get token system account")
+                    return jsonify({"status": False, "info": "You try get token system account"}), 400
+            else:
+                loger.error("URAccount", f"Name not in users")
+                return jsonify({"status": False, "info": "Name not in users"}), 404
 
         # change user token
         @app.route("/ChangeToken", methods=['POST'])
